+++ src/main/resources/tasks/cache-invalidate-all/cache-invalidate-all.ts
@@ -1,6 +1,6 @@
 import { logger } from '../../lib/utils/logging';
 import { CacheInvalidateAllConfig } from './cache-invalidate-all-config';
-import { frontendInvalidateAllSync } from '../../lib/cache/frontend-invalidate-requests';
+import { frontendInvalidateAllSync } from '../../lib/cache/frontend-cache';
 
 export const run = ({ retryIfFail, eventId }: CacheInvalidateAllConfig) => {
     logger.info(`Running task for full cache invalidation - eventId ${eventId}`);
